Move error to its own mod
