+++ src/lib.rs
@@ -7,6 +7,7 @@
 const F64_ASSERT_PRECISION: f64 = 1e-4f64;
 
 pub mod bounding_box;
+pub mod error;
 pub mod line_segment;
 pub mod point;
 pub mod polygon;
